refactor: #121 CMS レスポンスデータの型の持ち方を全体統一
