+++ src/content/articles/internals/remote.ts
@@ -1,58 +1,8 @@
-import type { ColorThemeBase } from "@content/commons";
-import {
-  fetchContents,
-  type MicroCMSFilters,
-  type MicroCMSImage,
-  type MicroCMSListContentBase,
-} from "@lib/utils/microcms";
+import { fetchContents, type MicroCMSFilters } from "@lib/utils/microcms";
 import { ensureNonNil } from "@lib/utils/type";
 import type { AstroIntegrationLogger } from "astro";
 import type { Article } from "../types";
-
-/**
- *
- */
-interface CMSArticle extends MicroCMSListContentBase {
-  /**
-   *
-   */
-  title: string;
-
-  /**
-   * comma-separated string
-   */
-  keywords?: string;
-
-  /**
-   *
-   */
-  description: string;
-
-  /**
-   *
-   */
-  "hero-image": MicroCMSImage;
-
-  /**
-   *
-   */
-  "hero-image-alt": string;
-
-  /**
-   *
-   */
-  "hero-iamge-label"?: string | null;
-
-  /**
-   *
-   */
-  "theme-color": [ColorThemeBase];
-
-  /**
-   *
-   */
-  content: string;
-}
+import type { CMSArticle } from "./cms-type";
 
 /**
  *
@@ -70,7 +20,7 @@ function convertCMSDataToArticle(cmsData: CMSArticle): Omit<Article, "id"> {
     content,
     "hero-image": heroImage,
     "hero-image-alt": heroImageAlt,
-    "hero-iamge-label": heroImageLabel,
+    "hero-image-label": heroImageLabel,
     "theme-color": themeColor,
   } = cmsData;
