+++ src/content/guidelines/index.ts
@@ -1,2 +1,3 @@
+export type * from "./internals/cms-type";
 export * from "./loader";
 export * from "./types";
