+++ src/content/meta/index.ts
@@ -1,3 +1,4 @@
 export * from "./getters";
+export type * from "./internals/cms-type";
 export * from "./loader";
 export * from "./types";
